+++ R/data.R
@@ -1,4 +1,4 @@
-#' Example Input for Mus Musculus Analysis (Myeloma)
+#' Example Input for Myelome Analysis (Mus Musculus)
 #'
 #' A dataset containing the differentially-expressed genes and adjusted p-values
 #' for the GEO dataset GSE99393. The RNA microarray experiment was perform to
@@ -8,10 +8,10 @@
 #' (myeloma-associated  macrophages), 3 of which were treated with CSF1R
 #' antibody (treatment group) and the rest were treated with control IgG
 #' antibody (control group). In this dataset, differentially-expressed genes
-#' with |logFC| >= 2 and FDR < 0.05 are presented.
-#' \emph{Generated on Oct 28, 2019.}
+#' with |logFC| >= 1 and FDR < 0.05 are presented.
+#' \emph{Generated on Nov 1, 2019.}
 #'
-#' @format A data frame with 45 rows and 2 variables: \describe{
+#' @format A data frame with 714 rows and 2 variables: \describe{
 #'   \item{Gene_Symbol}{MGI gene symbols of the differentially-expressed genes}
 #'   \item{FDR}{adjusted p values, via the Benjamini & Hochberg (1995) method}
 #' }
@@ -21,7 +21,7 @@
 #' \code{\link{run_pathfindR}} for details on the patfindR enrichment analysis.
 "myeloma_input"
 
-#' Example Output for Mus Musculus Analysis (Myeloma)
+#' Example Output for Myelome Analysis (Mus Musculus)
 #'
 #' A dataset containing the results of pathfindR's active-subnetwork-oriented
 #' enrichment workflow performed on the Mus musculus myeloma
@@ -280,7 +280,7 @@
 #'
 #' @format list containing 12345 vectors. Each vector is the set of gene symbols
 #' of interactors B for each interactor A.
-"gene_mania_adj_list"
+"genemania_adj_list"
 
 #' KEGG PIN Adjacency List
 #'
@@ -304,13 +304,26 @@
 #' of interactors B for each interactor A.
 "intact_adj_list"
 
+#' STRING PIN Adjacency List
+#'
+#' An adjacency list of vectors containing interactors B for each interactor A
+#' in the STRING protein-protein interaction network (The designations
+#' "interactor A" and "interactor B" are arbitrary). Only interactions with a combined
+#' score >= 800 were kept.
+#' \emph{Generated on Oct 31, 2019.}
+#'
+#' @format list containing 11934 vectors. Each vector is the set of gene symbols
+#' of interactors B for each interactor A.
+"string_adj_list"
+
 #' Mus musculus STRING PIN Adjacency List
 #'
 #' An adjacency list of vectors containing interactors B for each interactor A
 #' in the Mus musculus STRING protein-protein interaction network (The designations
-#' "interactor A" and "interactor B" are arbitrary).
-#' \emph{Generated on Oct 28, 2019.}
+#' "interactor A" and "interactor B" are arbitrary). Only interactions with a combined
+#' score >= 800 were kept.
+#' \emph{Generated on Oct 31, 2019.}
 #'
-#' @format list containing 19059 vectors. Each vector is the set of gene symbols
+#' @format list containing 11934 vectors. Each vector is the set of gene symbols
 #' of interactors B for each interactor A.
 "mmu_string_adj_list"