+++ NEWS.md
@@ -1,4 +1,4 @@
-# pathfindR 1.4.0.9010
+# pathfindR 1.4.0.9011
 ## To be released as 1.4.0
 
 ## Major Changes